+++ docs/changes.rst
@@ -24,6 +24,7 @@ Weblate 5.17.1
 * Attribute translations copied by :ref:`addon-weblate.autotranslate.autotranslate` to the add-on user.
 * Documented restoring Docker based setups from backups, see :ref:`restore-docker`.
 * :doc:`/admin/announcements` can now also be managed via the :ref:`api` for categories.
+* :ref:`addon-weblate.autotranslate.autotranslate` now records automatic translation results in the add-on activity log.
 
 .. rubric:: Bug fixes